+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/ParticleGenerator.scala
