+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/SufficientStatisticsVariableElimination.scala
@@ -16,7 +16,7 @@ package com.cra.figaro.algorithm.factored
 import com.cra.figaro.algorithm._
 import com.cra.figaro.algorithm.learning._
 import com.cra.figaro.language._
-import com.cra.figaro.library.factors._
+import com.cra.figaro.algorithm.factored.factors._
 import scala.collection._
 import scala.collection.mutable.{ Map, Set }
