+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/beliefpropagation/MPEBeliefPropagation.scala
@@ -17,7 +17,7 @@ import com.cra.figaro.algorithm._
 import com.cra.figaro.language._
 import com.cra.figaro.algorithm._
 import com.cra.figaro.algorithm.sampling._
-import com.cra.figaro.library.factors._
+import com.cra.figaro.algorithm.factored.factors._
 import com.cra.figaro.util
 import scala.collection.mutable.{ Set, Map }
