+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/factors/BasicFactor.scala
@@ -11,7 +11,7 @@
  * See http://www.github.com/p2t2/figaro for a copy of the software license.
  */
 
-package com.cra.figaro.library.factors
+package com.cra.figaro.algorithm.factored.factors
 
 import com.cra.figaro.util._
 import scala.annotation.tailrec
