+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/factors/Factor.scala
@@ -10,7 +10,7 @@
  * 
  * See http://www.github.com/p2t2/figaro for a copy of the software license.
  */
-package com.cra.figaro.library.factors
+package com.cra.figaro.algorithm.factored.factors
 
 import scala.annotation.tailrec
 import com.cra.figaro.language.Element
