+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/factors/factory/StarFactory.scala
@@ -11,10 +11,10 @@
  * See http://www.github.com/p2t2/figaro for a copy of the software license.
  */
 
-package com.cra.figaro.library.factors.factory
+package com.cra.figaro.algorithm.factored.factors.factory
 
 import com.cra.figaro.language._
-import com.cra.figaro.library.factors._
+import com.cra.figaro.algorithm.factored.factors._
 import com.cra.figaro.algorithm.lazyfactored._
 
 /**
