+++ Figaro/src/test/scala/com/cra/figaro/test/algorithm/AbstractionTest.scala
@@ -17,11 +17,11 @@ import org.scalatest.Matchers
 import org.scalatest.WordSpec
 import com.cra.figaro.algorithm._
 import com.cra.figaro.algorithm.factored._
+import com.cra.figaro.algorithm.factored.factors._
 import com.cra.figaro.algorithm.lazyfactored._
 import com.cra.figaro.language._
 import com.cra.figaro.library.atomic.continuous._
 import com.cra.figaro.library.compound._
-import com.cra.figaro.library.factors._
 import com.cra.figaro.test.tags.NonDeterministic
 
 class AbstractionTest extends WordSpec with Matchers {
