+++ Figaro/src/test/scala/com/cra/figaro/test/algorithm/factored/BPTest.scala
@@ -17,13 +17,13 @@ import scala.collection.mutable.Map
 import org.scalatest.WordSpec
 import org.scalatest.Matchers
 import com.cra.figaro.algorithm.factored._
+import com.cra.figaro.algorithm.factored.factors._
 import com.cra.figaro.algorithm.factored.beliefpropagation._
 import com.cra.figaro.language._
 import com.cra.figaro.library.compound.If
 import com.cra.figaro.library.atomic.discrete.{Uniform => DUniform}
 import com.cra.figaro.library.atomic.continuous.{Uniform => CUniform}
 import com.cra.figaro.library.compound.IntSelector
-import com.cra.figaro.library.factors._
 import com.cra.figaro.algorithm.lazyfactored.LazyValues
 import com.cra.figaro.algorithm.UnsupportedAlgorithmException
