+++ Figaro/src/test/scala/com/cra/figaro/test/algorithm/factored/FactorTest.scala
@@ -17,6 +17,7 @@ import org.scalatest.Matchers
 import org.scalatest.PrivateMethodTester
 import org.scalatest.WordSpec
 import com.cra.figaro.algorithm.Values
+import com.cra.figaro.algorithm.factored.factors._
 import com.cra.figaro.algorithm.lazyfactored.LazyValues
 import com.cra.figaro.algorithm.lazyfactored.Regular
 import com.cra.figaro.algorithm.lazyfactored.ValueSet
@@ -40,7 +41,6 @@ import com.cra.figaro.language.Universe
 import com.cra.figaro.library.atomic.continuous.Normal
 import com.cra.figaro.library.atomic.continuous.Uniform
 import com.cra.figaro.library.compound.CPD
-import com.cra.figaro.library.factors._
 import com.cra.figaro.algorithm.factored.ParticleGenerator
 
 class FactorTest extends WordSpec with Matchers with PrivateMethodTester {
