+++ Figaro/src/test/scala/com/cra/figaro/test/algorithm/sampling/SamplingUtil.scala
